requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
async-timeout==4.0.3
attrs==23.1.0
black==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
coverage==7.3.2
croniter==2.0.1
dill==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.0
flake8==6.1.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy==1.6.1
mypy-extensions==1.0.0
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.17.1
psutil==5.9.6
pycodestyle==2.11.1
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyflakes==3.1.0
pylint==3.0.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
Represent==1.6.0.post0
requests==2.31.0
rpds-py==0.12.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.3.24
sqlalchemy-aio==0.17.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
toolrack==3.0.1
types-jsonschema==4.19.0.4
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
zipp==3.17.0
